v2.0.0.
API for hook listeners has been changed. Hook listeners are passed in the original args, not the hook event. However the hook listeners are run in the context of the hook event, so they can access the , , , and  by referencing them via . Also changed behavior of how  is invoked.

Updated npm deps. Updated readme.